unrelated methods are moved out of interface and segregated to IBaseRepository.kt, IMetadataRepository.kt and ITrackListRepository.kt
