+++ feature_book_details/src/main/java/com/allsoftdroid/feature/book_details/data/repository/MetadataRepositoryImpl.kt
@@ -10,11 +10,12 @@ import com.allsoftdroid.database.metadataCacheDB.MetadataDao
 import com.allsoftdroid.feature.book_details.data.databaseExtension.SaveMetadataInDatabase
 import com.allsoftdroid.feature.book_details.data.databaseExtension.asMetadataDomainModel
 import com.allsoftdroid.feature.book_details.data.databaseExtension.asTrackDomainModel
+import com.allsoftdroid.feature.book_details.data.model.TrackFormat
 import com.allsoftdroid.feature.book_details.data.network.response.GetAudioBookMetadataResponse
 import com.allsoftdroid.feature.book_details.data.network.service.ArchiveMetadataService
 import com.allsoftdroid.feature.book_details.domain.model.AudioBookMetadataDomainModel
 import com.allsoftdroid.feature.book_details.domain.model.AudioBookTrackDomainModel
-import com.allsoftdroid.feature.book_details.domain.repository.AudioBookMetadataRepository
+import com.allsoftdroid.feature.book_details.domain.repository.IMetadataRepository
 import com.allsoftdroid.feature.book_details.utils.NetworkState
 import com.google.gson.Gson
 import kotlinx.coroutines.Dispatchers
@@ -26,12 +27,12 @@ import retrofit2.Callback
 import retrofit2.Response
 import timber.log.Timber
 
-class AudioBookMetadataRepositoryImpl(
-    private val metadataDao : MetadataDao,
+class MetadataRepositoryImpl(
+    metadataDao : MetadataDao,
     private val bookId: String,
     private val metadataDataSource : ArchiveMetadataService,
     private val saveInDatabase: SaveInDatabase<MetadataDao,SaveMetadataInDatabase>
-) : AudioBookMetadataRepository{
+) : IMetadataRepository{
 
     /**
      * Books type live data is fetched from the database and notify observer for any change in data.
@@ -46,50 +47,6 @@ class AudioBookMetadataRepositoryImpl(
 
     private val audioBookMetadata : LiveData<AudioBookMetadataDomainModel>
         get() = _audioBookMetadata
-
-    private val trackLoadEvent = MutableLiveData<Event<TrackFormat>>()
-
-    private var _audioBookTrackList : LiveData<List<AudioBookTrackDomainModel>>
-            = Transformations.switchMap(trackLoadEvent){
-        Transformations.map(
-            metadataDao.getTrackDetails(bookId,formatContains = "64")
-        ){
-            it.asTrackDomainModel()
-        }
-    }
-
-    private var _audioBookTrackList2 : LiveData<List<AudioBookTrackDomainModel>>
-            = Transformations.switchMap(trackLoadEvent){event ->
-        event.getContentIfNotHandled()?.let {format ->
-            when(format){
-                is TrackFormat.FormatBP64 -> {
-                    Transformations.map(
-                        metadataDao.getTrackDetails(bookId,formatContains = "64")
-                    ){
-                        it.asTrackDomainModel()
-                    }
-                }
-                is TrackFormat.FormatBP128 -> {
-                    Transformations.map(
-                        metadataDao.getTrackDetails(bookId,formatContains = "128")
-                    ){
-                        it.asTrackDomainModel()
-                    }
-                }
-                else ->{
-                    Transformations.map(
-                        metadataDao.getTrackDetailsVBR(bookId)
-                    ){
-                        it.asTrackDomainModel()
-                    }
-                }
-            }
-        }
-    }
-
-    private val audioBookTrackList : LiveData<List<AudioBookTrackDomainModel>>
-    get() = _audioBookTrackList2
-
 
     override fun getBookId() = bookId
 
@@ -138,13 +95,6 @@ class AudioBookMetadataRepositoryImpl(
     }
 
     override fun getMetadata() = audioBookMetadata
-
-    override suspend fun loadTrackListData(format: TrackFormat) {
-        trackLoadEvent.value = Event(format)
-        Timber.d("Track format :${format}")
-    }
-
-    override fun getTrackList() = audioBookTrackList
 
     override fun networkResponse(): Variable<Event<NetworkState>> = _networkResponse
 }
